+++ lib/service/failure_report.go
@@ -11,7 +11,7 @@ import (
 
 type FailureReportAPI interface {
 	RegisterFailureReport(failure *asset.NewFailureReport, owner string) (*asset.FailureReport, error)
-	GetFailureReport(computeTaskKey string) (*asset.FailureReport, error)
+	GetFailureReport(assetKey string) (*asset.FailureReport, error)
 }
 
 type FailureReportServiceProvider interface {
@@ -22,6 +22,7 @@ type FailureReportDependencyProvider interface {
 	LoggerProvider
 	persistence.FailureReportDBALProvider
 	ComputeTaskServiceProvider
+	FunctionServiceProvider
 	EventServiceProvider
 	TimeServiceProvider
 }
@@ -41,31 +42,26 @@ func (s *FailureReportService) RegisterFailureReport(newFailureReport *asset.New
 	if err != nil {
 		return nil, errors.FromValidationError(asset.FailureReportKind, err)
 	}
-
-	task, err := s.GetComputeTaskService().GetTask(newFailureReport.ComputeTaskKey)
-	if err != nil {
-		return nil, err
+	switch newFailureReport.AssetType {
+	case asset.FailedAssetKind_FAILED_ASSET_COMPUTE_TASK:
+		err = s.processTaskFailure(newFailureReport.AssetKey, requester)
+	case asset.FailedAssetKind_FAILED_ASSET_FUNCTION:
+		err = s.processFunctionFailure(newFailureReport.AssetKey, requester)
+	default:
+		return nil, errors.NewBadRequest("can only register failure for asset_kind values function and compute task")
 	}
 
-	if task.Worker != requester {
-		return nil, errors.NewPermissionDenied(fmt.Sprintf("only %q worker can register failure report", task.Worker))
-	}
-
-	if task.Status != asset.ComputeTaskStatus_STATUS_DOING {
-		return nil, errors.NewBadRequest(fmt.Sprintf("cannot register failure report for task with status %q", task.Status.String()))
-	}
-
-	err = s.GetComputeTaskService().ApplyTaskAction(task.Key, asset.ComputeTaskAction_TASK_ACTION_FAILED, "failure report registered", requester)
 	if err != nil {
 		return nil, err
 	}
 
 	failureReport := &asset.FailureReport{
-		ComputeTaskKey: newFailureReport.ComputeTaskKey,
-		ErrorType:      newFailureReport.ErrorType,
-		LogsAddress:    newFailureReport.LogsAddress,
-		CreationDate:   timestamppb.New(s.GetTimeService().GetTransactionTime()),
-		Owner:          requester,
+		AssetKey:     newFailureReport.AssetKey,
+		AssetType:    newFailureReport.AssetType,
+		ErrorType:    newFailureReport.ErrorType,
+		LogsAddress:  newFailureReport.LogsAddress,
+		CreationDate: timestamppb.New(s.GetTimeService().GetTransactionTime()),
+		Owner:        requester,
 	}
 
 	err = s.GetFailureReportDBAL().AddFailureReport(failureReport)
@@ -75,7 +71,7 @@ func (s *FailureReportService) RegisterFailureReport(newFailureReport *asset.New
 
 	event := &asset.Event{
 		EventKind: asset.EventKind_EVENT_ASSET_CREATED,
-		AssetKey:  failureReport.ComputeTaskKey,
+		AssetKey:  failureReport.AssetKey,
 		AssetKind: asset.AssetKind_ASSET_FAILURE_REPORT,
 		Asset:     &asset.Event_FailureReport{FailureReport: failureReport},
 	}
@@ -87,7 +83,56 @@ func (s *FailureReportService) RegisterFailureReport(newFailureReport *asset.New
 	return failureReport, nil
 }
 
-func (s *FailureReportService) GetFailureReport(computeTaskKey string) (*asset.FailureReport, error) {
-	s.GetLogger().Debug().Str("computeTaskKey", computeTaskKey).Msg("Get failure report")
-	return s.GetFailureReportDBAL().GetFailureReport(computeTaskKey)
+func (s *FailureReportService) GetFailureReport(assetKey string) (*asset.FailureReport, error) {
+	s.GetLogger().Debug().Str("assetKey", assetKey).Msg("Get failure report")
+	return s.GetFailureReportDBAL().GetFailureReport(assetKey)
+}
+
+func checkTaskPermissions(task *asset.ComputeTask, requester string) error {
+	if task.Worker != requester {
+		return errors.NewPermissionDenied(fmt.Sprintf("only %q worker can register failure report for compute task", task.Worker))
+	}
+
+	if task.Status != asset.ComputeTaskStatus_STATUS_DOING {
+		return errors.NewBadRequest(fmt.Sprintf("cannot register failure report for task with status %q", task.Status.String()))
+	}
+
+	return nil
+}
+
+func (s *FailureReportService) processTaskFailure(taskKey string, requester string) error {
+	task, err := s.GetComputeTaskService().GetTask(taskKey)
+	if err != nil {
+		return err
+	}
+
+	err = checkTaskPermissions(task, requester)
+	if err != nil {
+		return err
+	}
+
+	return s.GetComputeTaskService().ApplyTaskAction(taskKey, asset.ComputeTaskAction_TASK_ACTION_FAILED, "failure report registered", requester)
+}
+
+func checkFunctionPermissions(function *asset.Function, requester string) error {
+	if function.Owner != requester {
+		return errors.NewPermissionDenied(fmt.Sprintf("only %q owner can register failure report for function", function.Owner))
+	}
+
+	return nil
+}
+
+func (s *FailureReportService) processFunctionFailure(functionKey string, requester string) error {
+	function, err := s.GetFunctionService().GetFunction(functionKey)
+	if err != nil {
+		return err
+	}
+
+	err = checkFunctionPermissions(function, requester)
+
+	if err != nil {
+		return err
+	}
+
+	return s.GetFunctionService().ApplyFunctionAction(functionKey, asset.FunctionAction_FUNCTION_ACTION_FAILED, "failure report registered", requester)
 }