+++ server/standalone/dbal/failure_report_test.go
@@ -3,13 +3,12 @@ package dbal
 import (
 	"context"
 	"errors"
-	"testing"
-
 	"github.com/jackc/pgx/v4"
 	"github.com/pashagolub/pgxmock"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	orcerrors "github.com/substra/orchestrator/lib/errors"
+	"testing"
 )
 
 func TestGetFailureReportNotFound(t *testing.T) {
@@ -18,9 +17,9 @@ func TestGetFailureReportNotFound(t *testing.T) {
 
 	mock.ExpectBegin()
 
-	computeTaskKey := "4c67ad88-309a-48b4-8bc4-c2e2c1a87a83"
+	assetKey := "4c67ad88-309a-48b4-8bc4-c2e2c1a87a83"
 	mock.ExpectQuery(`SELECT .* FROM expanded_failure_reports`).
-		WithArgs(testChannel, computeTaskKey).
+		WithArgs(assetKey, testChannel).
 		WillReturnError(pgx.ErrNoRows)
 
 	tx, err := mock.Begin(context.Background())
@@ -28,7 +27,7 @@ func TestGetFailureReportNotFound(t *testing.T) {
 
 	dbal := &DBAL{ctx: context.TODO(), tx: tx, channel: testChannel}
 
-	_, err = dbal.GetFailureReport(computeTaskKey)
+	_, err = dbal.GetFailureReport(assetKey)
 
 	assert.Error(t, err)
 	orcError := new(orcerrors.OrcError)